início da modelagem de dados
