+++ projetoNovo/Model/Model.php
@@ -1,0 +1,10 @@
+<?php
+	namespace projetoNovo\Model;
+	class Model{
+		protected $db;
+
+		public function __construct(\PDO $db){
+			$this->db = $db;
+		}
+	}
+ ?>
