--- projetoNovo/Model/usuario.php
@@ -1,21 +1,0 @@
-<?php
-    class Usuario{
-        private $matricula;
-        private $email;
-        private $ctt;
-        private $id;
-        private $nick;
-        private $nome;
-        private $email_corp;
-        private $ctt_corp;
-        private $senha;
-        private $rank;
-        private $Matricula;
-            public function solicitarCadastro(){
-                
-            }
-            public function novoCadastro(){
-                
-            }
-    }
-?>
