+++ projetoNovo/View/Admin/index.php
@@ -1,0 +1,1 @@
+teste
